requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.29
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
pyjwt>=2.8.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-multipart>=0.0.9
httpx>=0.27.0
requests>=2.31.0

# Frontend (use newer streamlit compatible with Python 3.13)
streamlit>=1.40.0
extra-streamlit-components>=0.1.71

# Hardware
pyserial>=3.5
python-escpos>=3.1

# Utilities (numpy 2.x is fine, no C compiler needed with pre-built wheels)
numpy>=1.26.0
pandas>=2.0.0
plotly>=5.20.0
